Fixing "String RoboteqSerial::readQuery" for RX availability
Method "String RoboteqSerial::readQuery(const char *message)" expects, on top of the timeout, that RX is available. That's OK, but in cases where the processor used is fast (like ESP32) the Roboteq controller may have not replied yet when readQuery is called after a "sendQuery". So, the _stream.available() should be inside the while.